Made it my own
I don't use zsh or many of the tools included. As such I had to rewrite
the zshrc.symlink file for bash. I decided to rip out a lot of the
functionality that the original file had to run certain scripts before
others. I figure, if I need to do that, I can prefix the scripts with
00-99 much like old school init scripts, and let bash handle sorting
that out. I also added Xresources to make urxvt not look like shite.